Remove ClientReader::maybe_consensus_state()
